requirements.txt
django==1.11
djangorestframework
simple_salesforce
